Move config to a separate package A package needing the Config struct, while also getting started by Start(), will create a circular dependency. In this diff, I separate the config logic to have any package import it.
